pyproject.toml
[project]
name = "master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Pranav Murthy", email = "pmurthy2019@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.2.0",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
]

[project.scripts]
master = "master:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
